argocd: update to 1.8.3.3
argocd-appilcation-controller is changed from Deployment
to StatefulSet.  Because of this, we need to upgrade
argocd manually.

Starting from v1.8, argocd no longer check the health
status of Applications.  Although the upstream provides
a custom Lua script to re-enable the feature, it causes
error like this:

    ComparisonError: json: cannot unmarshal array into Go value of type health.HealthStatus

Therefore, we simply do not rely on the health check of
Applications.

These tools are also updated:

- dex 2.27.0.2
- redis 6.0.10.1
- kustomize 3.9.2